Unify formatting datetime in orders & transactions & remove duplicate code
